apps/api/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
httpx==0.27.2
presidio-analyzer==2.2.355
presidio-anonymizer==2.2.355
spacy==3.7.6
faker==30.8.2
tiktoken==0.8.0
pyjwt==2.9.0
stripe==10.12.0
supabase==2.9.1
readability-lxml==0.8.1
lxml[html_clean]==5.3.0
sse-starlette==2.1.3
python-multipart==0.0.12
cryptography==43.0.1
